PCI: qcom: Add support for host_stop_link() & host_start_link()
For the switches like QPS615 which needs to configure it before
the PCIe link is established.

if the link is not up assert the PERST# and disable LTSSM bit so
that PCIe controller will not participate in the link training
as part of host_stop_link().

De-assert the PERST# and enable LTSSM bit back in host_start_link().

Introduce ltssm_disable function op to stop the link training.

Signed-off-by: Krishna chaitanya chundru <quic_krichai@quicinc.com>
Link: https://lore.kernel.org/r/20240803-qps615-v2-7-9560b7c71369@quicinc.com