of: overlays: dwc3-flattening: Provide overlay symbols
The overlays introduced to apply the migration to the flattened Qualcomm
DWC3 binding depends on symbols to be applied, but in general the
DeviceTree blobs in circulation on people's devices wasn't built with
symbols - and hence they can not be applied.

Work around this problem by detecting when the live DeviceTree is
lacking symbols and inject the necessary symbols for each target, for
the duration of the dwc3-flattening overlay application.

Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>